- Added Programmable Interrupt Controller (PIC) device driver - And other minor updates.
git-svn-id: svn://localhost/mosa@135 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Mosa/DeviceDrivers/PCI/IPCIController.cs
@@ -7,12 +7,31 @@
  *  Phil Garcia (tgiphil) <phil@thinkedge.com>
  */
 
-
 namespace Mosa.DeviceDrivers.PCI
 {
+	/// <summary>
+	/// Interface to a PCI Controller
+	/// </summary>
 	public interface IPCIController
 	{
-		uint ReadConfig(uint bus, uint slot, uint function, uint register);
-		void WriteConfig(uint bus, uint slot, uint function, uint register, uint value);
+		/// <summary>
+		/// Reads from configuraton space
+		/// </summary>
+		/// <param name="bus">The bus.</param>
+		/// <param name="slot">The slot.</param>
+		/// <param name="function">The function.</param>
+		/// <param name="register">The register.</param>
+		/// <returns></returns>
+		uint ReadConfig(byte bus, byte slot, byte function, byte register);
+		
+		/// <summary>
+		/// Writes to configuraton space
+		/// </summary>
+		/// <param name="bus">The bus.</param>
+		/// <param name="slot">The slot.</param>
+		/// <param name="function">The function.</param>
+		/// <param name="register">The register.</param>
+		/// <param name="value">The value.</param>
+		void WriteConfig(byte bus, byte slot, byte function, byte register, uint value);
 	}
 }
